try to get gunMat and screw also recorded.
those two piece of info is only present in code we have to get them
manually extracted. moreover, we should just go with no guessing at
all - the asset is not some large chunk of data and the point of
having this lookup table is that we want to avoid having to sync with
game client on those garbage.